requirements.txt
pydrive 
pandas 
google-auth 
google-auth-oauthlib
python-dotenv
duckdb
google-cloud-storage
gcsfs